// src/app/interfaces/users.ts
/**
 * Initialize UserTable
 */
export interface UserTable {
  /**
   * Setup of type string
   */
  createdAt: string

  /**
   * Setup of type string
   */
  name: string

  /**
   * Setup of type string
   */
  address: string

  /**
   * Setup of type string
   */
  image: string

  /**
   * Setup of type string
   */
  email: string

  /**
   * Setup of type string
   */
  phone: string

  /**
   * Setup of type string
   */
  jobTitle: string

  /**
   * Setup of type string
   */
  id: string
}